Moved EntityBase from Component.h and Component.cpp
