+++ esphome/components/esp32_camera/esp32_camera.h
@@ -3,6 +3,7 @@
 #ifdef USE_ESP32
 
 #include "esphome/core/component.h"
+#include "esphome/core/entity_base.h"
 #include "esphome/core/helpers.h"
 #include <esp_camera.h>
 #include <freertos/FreeRTOS.h>
