+++ esphome/components/select/select.h
@@ -3,6 +3,7 @@
 #include <set>
 #include <utility>
 #include "esphome/core/component.h"
+#include "esphome/core/entity_base.h"
 #include "esphome/core/helpers.h"
 
 namespace esphome {
