+++ esphome/components/sensor/sensor.h
@@ -2,6 +2,7 @@
 
 #include "esphome/core/log.h"
 #include "esphome/core/component.h"
+#include "esphome/core/entity_base.h"
 #include "esphome/core/helpers.h"
 #include "esphome/components/sensor/filter.h"
