+++ esphome/components/switch/switch.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "esphome/core/component.h"
+#include "esphome/core/entity_base.h"
 #include "esphome/core/preferences.h"
 #include "esphome/core/helpers.h"
